moved state and fetch from components to app.js, did a post fetch i app
