+++ code/src/components/MessageList.js
@@ -1,28 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import moment from 'moment';
 
-import { API_URL } from '../reusable/urls'
-
-
-const MessageList = () => {
-
-  const [MessageList, setMessageList] = useState([])
-  console.log(MessageList)
-
-  useEffect(() => {
-    fetchMessageList();
-  }, [])
-
-const fetchMessageList = () => {
-  fetch(API_URL)
-  .then(res => res.json())
-  .then(messages => setMessageList(messages))
-  .catch(err => console.log(err));
-}
+const MessageList = ({MessageListArray}) => {
+  console.log(MessageListArray);
 
   return (
     <div className="message-list-container">
-      {MessageList.map(message => (
+      {MessageListArray.map(message => (
         <div className= "thoughts-card" key={message._id}>
           <h4 className="message-text">{message.message}</h4>
         <div className="created-and-likes">
