+++ code/src/index.css
@@ -73,6 +73,7 @@ code {
 font-family: "Inconsolata",monospace;
 font-weight: 300;
 font-size: 16px;
+overflow-wrap: break-word;
 }
 
 .created{
